requirements.txt
streamlit==1.25.0
tensorflow==2.17.0
numpy==1.26.4
requests==2.31.0
pillow==9.4.0